[embolden] Rename to hb-outline
